TST: Separated pipeline tests
